requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-RESTful==0.3.10
Flask-Validator==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
pymongo==4.5.0
dnspython==2.4.2
marshmallow==3.20.1
cerberus==1.3.4
python-dateutil==2.8.2
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0 